Implementing Default Methods for Repositories.
Since I'm not expecting multiple implementations for the different repository types, I've gone ahead and moved the implementation details into their respective interfaces. This also allows for the UserRepository interface to extend both, so we don't have to worry about a tall-but-narrow hierarchy, but can have a wide-and-flat one instead. This means if we find the need to project results into a DTO, we just modify the interface and there it is, instead of having to ensure our inheritance structure is correct.