Beginning to test the User Controller.
This required an update to ignore the preferences list in the "isEqualTo" method for the User Object.
Also, the removal of the ExecutorService object in the controller itself. The async returns were too quick during testing and we were getting empty results during the test. This also allowed us to get rid of the searching and saving flags, as we'll just let JSF do the waiting.