User Controller Testing has been expanded.
I'm not one to say that this is complete until we get the UI up and running, but it captures the basic idea of how I expect the controller to work.

It's been interesting to work from the back->forward with regards to the setup of this app. Should be fun designing the GUI around what we've already built here. It helps a LOT already knowing how it generally works though.

Lot's of stuff to do still though: Profile and Preference Management, along with Roles and Permission controls.